Todas as exceçoes tratadas

+++ src/exceptions/MatriculaDiscenteRepetidaException.java
@@ -1,7 +1,7 @@
 package exceptions;
 
 public class MatriculaDiscenteRepetidaException extends Exception {
-	public MatriculaDiscenteRepetidaException (int codigo) {
+	public MatriculaDiscenteRepetidaException (long codigo) {
 		System.out.println("Código repetido para discente: " + codigo + ".");
 	}
